scripts/db/requirements.txt
psycopg2-binary==2.9.7
pandas==2.1.1
redis==4.6.0
python-dotenv==1.0.0
